complete folder restructure
