+++ dnas/bridge/workdir/dna.yaml
@@ -14,6 +14,10 @@ integrity:
       hash: ~
       bundled: "../../../target/wasm32-unknown-unknown/release/profiles_integrity.wasm"
       dependencies: ~
+    - name: transaction_requests_integrity
+      hash: ~
+      bundled: "../../../target/wasm32-unknown-unknown/release/transaction_requests_integrity.wasm"
+      dependencies: ~
 coordinator:
   zomes:
     - name: bridge
@@ -30,4 +34,4 @@ coordinator:
       hash: ~
       bundled: "../../../target/wasm32-unknown-unknown/release/transaction_requests.wasm"
       dependencies:
-        - name: profiles_integrity
+        - name: transaction_requests_integrity
