+++ dnas/bridge/zomes/integrity/bridge/src/lib.rs
@@ -4,25 +4,19 @@ pub use authority_list::*;
 pub mod properties;
 pub use properties::*;
 
-pub mod transaction;
-pub use transaction::*;
-
 mod validation;
 
-use types::*;
+use types::{ AuthorityList };
 use hdi::prelude::*;
 
 #[derive(Serialize, Deserialize)]
-#[serde(tag = "type")]
 #[hdk_entry_defs]
 #[unit_enum(UnitEntryTypes)]
 pub enum EntryTypes {
-    AuthorityList(AuthorityList),
-    Transaction(Transaction),
-    TransactionRequest(TransactionRequest),
+    AuthorityList(AuthorityList)
 }
 
 #[hdk_link_types]
 pub enum LinkTypes {
-    AgentPubKeyToTransactionRequest
+    AgentPubKeyToAuthorityList
 }
