+++ dnas/types/src/lib.rs
@@ -169,21 +169,3 @@ pub struct CreateTransactionRequestInput {
 
 #[derive(Debug, Serialize, Deserialize, SerializedBytes, Clone, PartialEq)]
 pub struct UIEnum(pub String);
-
-#[derive(Serialize, Deserialize, Debug)]
-pub enum SignalType {
-    TransactionRequestReceived {
-        transaction_request_hash: ActionHashB64,
-        transaction_request: TransactionRequest,
-    },
-    TransactionRequestAccepted {
-        transaction_request_hash: ActionHashB64,
-        transaction: (ActionHashB64, Transaction),
-    },
-    TransactionRequestCancelled {
-        transaction_request_hash: ActionHashB64,
-    },
-    TransactionRequestRejected {
-        transaction_request_hash: ActionHashB64,
-    },
-}
